+++ src/main/java/net/lugo/overlaylib/renderers/SimpleTextureOverlayRenderer.java
@@ -10,8 +10,8 @@ import net.minecraft.resources.Identifier;
 public class SimpleTextureOverlayRenderer extends OverlayRenderer {
     private static final float EPSILON = 1E-3f;
 
-    public SimpleTextureOverlayRenderer(Identifier texture, boolean doIrisFlickerFix, double nearbyBlockDistanceSq) {
-        super(RenderPipelines.POSITION_TEX_COLOR_FOG_TRIANGLES, texture, doIrisFlickerFix, nearbyBlockDistanceSq);
+    public SimpleTextureOverlayRenderer(Identifier texture) {
+        super(RenderPipelines.POSITION_TEX_COLOR_FOG_TRIANGLES, texture);
     }
 
     @Override